requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Web Framework
flask>=2.0.0
flask-restful>=0.3.9

# Data Processing
joblib>=1.0.0
pyyaml>=5.4.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Utilities
python-dotenv>=0.19.0

# Optional (comment out if not needed)
# pyjwt>=2.3.0  # Uncomment for JWT support
# cryptography>=3.4.0